requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
alembic==1.13.2
psycopg[binary]==3.2.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
python-multipart==0.0.9
